Switch to the vagrant provider.
The docker provider does not work well with hosted drone.io
services. It likely will work great with an internally hosted
open source drone server, which is docker based. However,
drone.io as it exists today is not actually docker backed which
prevents us from being able to specify the docker box to use
for building. This will be changed at some point in the future.